+++ CHANGES.md
@@ -9,6 +9,7 @@ Version 0.3.0
 To be released.
 
  -  Added `parseLogLevel()` function.
+ -  Added `isLogLevel()` function.
 
 
 Version 0.2.2